+++ CHANGELOG.md
@@ -12,6 +12,9 @@ All notable changes to this project will be documented in this file.
  - `google` keyword to `package.json` 
  - New `Parameters` section in configuration
  - Added `Online` and `Local` options for all nodes (`tf coco ssd` is the only supported for now)
+ - `.eslintrc.js` file
+ - Override of messages properties (threshold and max. detections)
+ - Two functinos to filter and count classes
 
 ### Changed
  - General polish `JavaScript` code across all nodes
@@ -28,13 +31,13 @@ All notable changes to this project will be documented in this file.
  - Set all used libraries to `const` variables
  - `Min. Threshold` and `Max. Detections` are now selected from an easier spinner
  - `Min. Threshold` is now defined from 0 to 100 in %
+ - Refactor of the variable types whether to use `var`, `let` or `const` 
 
 ### Fixed
  - Correct model initial loading for `tf coco ssd` node 
 
 ### Removed
  - Dependancies from `express` and `compression` npm modules
- - Override of messages properties - will come back on a future release
 
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),